- Refactored the code to remove unused arguments and comments - Added docs at necessary places - Removed matrix_stack from skia Style class - Moved push_matrix, pop_matrix and style2D class to openglrenderer
