+++ p5/sketch/Vispy2DRenderer/openglrenderer.py
@@ -219,11 +219,23 @@ def get_render_primitives(shape):
             render_primitives.extend(_get_borders(shape))
     return render_primitives
 
+@dataclass
+class Style2D:
+    background_color = (0.8, 0.8, 0.8, 1.0)
+    fill_color = COLOR_WHITE
+    fill_enabled = True
+    stroke_color = COLOR_BLACK
+    stroke_enabled = True
+    stroke_weight = 1
+    tint_color = COLOR_BLACK
+    tint_enabled = False
+    ellipse_mode = "CENTER"
+    rect_mode = "CORNER"
+    color_parse_mode = "RGB"
+    color_range = (255, 255, 255, 255)
 
 
 # Abstract class that contains common code for OpenGL renderers
-
-
 class OpenGLRenderer(ABC):
     def __init__(self, src_fbuffer, src_default):
         self.fbuffer_prog = Program(src_fbuffer.vert, src_fbuffer.frag)
@@ -269,6 +281,9 @@ class OpenGLRenderer(ABC):
         # Renderer Globals: RENDERING
         self.draw_queue = []
 
+        self.style = Style2D()
+        self.style_stack = []
+        self.matrix_stack = []
 
     def render_default(self, draw_type, draw_queue):
         # 1. Get the maximum number of vertices persent in the shapes
@@ -336,3 +351,14 @@ class OpenGLRenderer(ABC):
         # need np.newaxis to broadcast the vector because each row represents a vertex
         dehomogenized = product / product[:,3][:,np.newaxis]
         return dehomogenized[:,:3] # Return the first three rows of the result
+
+    def push_matrix(self):
+        """Pushes the current transformation matrix onto the matrix stack.
+        """
+        self.matrix_stack.append(self.transform_matrix.copy())
+
+    def pop_matrix(self):
+        """Pops the current transformation matrix off the matrix stack.
+        """
+        assert len(self.matrix_stack) > 0, "No matrix to pop"
+        self.transform_matrix = self.matrix_stack.pop()
