+++ p5/sketch/events.py
@@ -137,7 +137,12 @@ class Event:
     def __init__(self, raw_event, active=False):
         self._modifiers = list(map(lambda k: k.name, raw_event.modifiers))
         self._active = active
+
         self._raw = raw_event
+        """
+        _raw: VispyEvent, SkiaPseudoEvent
+            This holds the raw event generated by the backend 
+        """
 
     @property
     def modifiers(self):
@@ -191,7 +196,7 @@ class KeyEvent(Event):
     """Encapsulates information about a key event.
 
     :param key: The key associated with this event.
-    :type key: str
+    :type key: str | Key
 
     :param pressed: Specifies whether the key is held down or not.
     :type pressed: bool
